selectors folder renamed to "reducers"
