+++ assignment2/src/components/Dashboard.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { getTopGainersLosers } from '../store/selectors/stocks.js'; 
+import { getTopGainersLosers } from '../store/reducers/stocks.js'; 
 import { gainersSelector, losersSelector } from '../customSelectors/StocksSelectors.js';
 import GainersTable from './GainersTable';
 import LosersTable from './LosersTable';
