+++ assignment2/src/store/reducers/stocks.js
